requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.3
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-filebrowser-no-grappelli==3.7.7
django-heroku==0.3.1
django-js-asset==1.2.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.10
Faker==1.0.0
gunicorn==19.9.0
idna==2.7
mysqlclient==1.3.13
oauthlib==2.1.0
Pillow==5.3.0
psycopg2==2.7.7
PyJWT==1.6.4
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
text-unidecode==1.2
urllib3==1.24.1
whitenoise==4.1.2
